requirements.txt
pdfplumber==0.10.3
flask==3.0.0
flask-cors==4.0.0
jinja2==3.1.2
mlflow==2.9.2
colorlog==6.8.0
python-dotenv==1.0.0
requests==2.31.0
